Restructure files/classes
